refactor(@angular/ssr): export tokens from the `/tokens` entry point
This commit relocates the DI tokens to a dedicated entry point. This change ensures that we do not depend on tree-shaking to eliminate server code from client bundles.

Closes #28670
