+++ goldens/public-api/angular/ssr/index.api.md
@@ -5,7 +5,6 @@
 ```ts
 
 import { EnvironmentProviders } from '@angular/core';
-import { InjectionToken } from '@angular/core';
 
 // @public
 export class AngularAppEngine {
@@ -34,15 +33,6 @@ export enum RenderMode {
     Prerender = 3,
     Server = 1
 }
-
-// @public
-export const REQUEST: InjectionToken<Request>;
-
-// @public
-export const REQUEST_CONTEXT: InjectionToken<unknown>;
-
-// @public
-export const RESPONSE_INIT: InjectionToken<ResponseInit>;
 
 // @public
 export type ServerRoute = ServerRouteAppShell | ServerRouteClient | ServerRoutePrerender | ServerRoutePrerenderWithParams | ServerRouteServer;
