+++ goldens/public-api/angular/ssr/tokens/index.api.md
@@ -1,0 +1,20 @@
+## API Report File for "@angular/ssr_tokens"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { InjectionToken } from '@angular/core';
+
+// @public
+export const REQUEST: InjectionToken<Request>;
+
+// @public
+export const REQUEST_CONTEXT: InjectionToken<unknown>;
+
+// @public
+export const RESPONSE_INIT: InjectionToken<ResponseInit>;
+
+// (No @packageDocumentation comment for this package)
+
+```
