+++ packages/angular/ssr/tokens/BUILD.bazel
@@ -1,0 +1,19 @@
+load("//tools:defaults.bzl", "ts_library")
+
+package(default_visibility = ["//visibility:public"])
+
+ts_library(
+    name = "tokens",
+    srcs = glob(
+        [
+            "*.ts",
+            "src/**/*.ts",
+        ],
+    ),
+    module_name = "@angular/ssr/tokens",
+    tsconfig = "//:tsconfig-build-ng",
+    deps = [
+        "@npm//@angular/core",
+        "@npm//tslib",
+    ],
+)
